add scoring utils
